# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: eBay Price Checker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /spreadsheet/config:
    get:
      operationId: getSpreadsheetConfig
      tags: [spreadsheet]
      summary: Get current spreadsheet configuration
      responses:
        "200":
          description: Spreadsheet config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpreadsheetConfig"
    post:
      operationId: saveSpreadsheetConfig
      tags: [spreadsheet]
      summary: Save spreadsheet configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpreadsheetConfigRequest"
      responses:
        "200":
          description: Config saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpreadsheetConfig"
...
  /spreadsheet/row/{row}:
    get:
      operationId: getSpreadsheetRow
      tags: [spreadsheet]
      summary: Read a row from spreadsheet to get my price
      parameters:
        - name: row
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Row data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpreadsheetRowData"
        "400":
          description: Not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    SpreadsheetConfigRequest:
      type: object
      properties:
        spreadsheetId:
          type: string
        sheetName:
          type: string
        alertColumnIndex:
          type: integer
          description: Column index (0-based) for price alert
        priceColumnIndex:
          type: integer
          description: Column index (0-based) for competitor price write-back
        myPriceColumnIndex:
          type: integer
          description: Column index (0-based) where my own price is stored in the sheet
        sourceUrlColumnIndex:
          type: integer
        ebayUrlColumnIndex:
          type: integer
        lowestPriceColumnIndex:
          type: integer
        lowestConditionColumnIndex:
          type: integer
        lastCheckColumnIndex:
          type: integer
        alertStatusColumnIndex:
          type: integer
        repricedValueColumnIndex:
          type: integer
        evidenceUrlsColumnIndex:
          type: integer
        trackedTargetUrlColumnIndex:
          type: integer
        trackedTargetConditionColumnIndex:
          type: integer
        trackedTargetPriceColumnIndex:
          type: integer
        autoRepriceEnabled:
          type: boolean
        undercutAmount:
          type: number
        minAllowedPrice:
          type: number
        ebayAppId:
          type: string
        ebayDevId:
          type: string
        ebayCertId:
          type: string
        ebayUserToken:
          type: string
        inventoryCheckerBaseUrl:
          type: string
        inventoryCheckerApiKey:
          type: string
        inventoryStatusColumnIndex:
          type: integer
        ebayListingConditionColumnIndex:
          type: integer
          nullable: true
        ebayOAuthRedirectUri:
          type: string
        ebayOAuthClientSecret:
          type: string
        serviceAccountJson:
          type: string
          description: Google Service Account JSON (stored securely)
      required:
        - spreadsheetId
        - sheetName
    SpreadsheetConfig:
      type: object
      properties:
        spreadsheetId:
          type: string
        sheetName:
          type: string
        alertColumnIndex:
          type: integer
        priceColumnIndex:
          type: integer
        myPriceColumnIndex:
          type: integer
        sourceUrlColumnIndex:
          type: integer
        ebayUrlColumnIndex:
          type: integer
        lowestPriceColumnIndex:
          type: integer
        lowestConditionColumnIndex:
          type: integer
        lastCheckColumnIndex:
          type: integer
        alertStatusColumnIndex:
          type: integer
        repricedValueColumnIndex:
          type: integer
        evidenceUrlsColumnIndex:
          type: integer
        trackedTargetUrlColumnIndex:
          type: integer
        trackedTargetConditionColumnIndex:
          type: integer
        trackedTargetPriceColumnIndex:
          type: integer
        autoRepriceEnabled:
          type: boolean
        undercutAmount:
          type: number
        minAllowedPrice:
          type: number
        hasEbayCredentials:
          type: boolean
        hasEbayOAuth:
          type: boolean
        inventoryCheckerBaseUrl:
          type: string
        hasInventoryCheckerApiKey:
          type: boolean
        inventoryStatusColumnIndex:
          type: integer
        ebayListingConditionColumnIndex:
          type: integer
        ebayOAuthRedirectUri:
          type: string
        isConfigured:
          type: boolean
        hasServiceAccount:
          type: boolean
        databaseNeedsMigration:
          type: boolean
          description: True when Postgres schema is missing (run drizzle-kit push)
        setupMessageJa:
          type: string
          description: User-facing hint in Japanese
      required:
        - isConfigured
        - hasServiceAccount
    SpreadsheetRowData:
      type: object
      properties:
        rowNumber:
          type: integer
        myPrice:
          type: number
        rawValue:
          type: string
        found:
          type: boolean
      required:
        - rowNumber
        - found
...