# lib/api-zod/src/generated/types/spreadsheetConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * eBay Price Checker API
 * OpenAPI spec version: 0.1.0
 */

export interface SpreadsheetConfig {
  spreadsheetId?: string;
  sheetName?: string;
  alertColumnIndex?: number;
  priceColumnIndex?: number;
  myPriceColumnIndex?: number;
  sourceUrlColumnIndex?: number;
  ebayUrlColumnIndex?: number;
  lowestPriceColumnIndex?: number;
  lowestConditionColumnIndex?: number;
  lastCheckColumnIndex?: number;
  alertStatusColumnIndex?: number;
  repricedValueColumnIndex?: number;
  evidenceUrlsColumnIndex?: number;
  trackedTargetUrlColumnIndex?: number;
  trackedTargetConditionColumnIndex?: number;
  trackedTargetPriceColumnIndex?: number;
  autoRepriceEnabled?: boolean;
  undercutAmount?: number;
  minAllowedPrice?: number;
  hasEbayCredentials?: boolean;
  hasEbayOAuth?: boolean;
  inventoryCheckerBaseUrl?: string;
  hasInventoryCheckerApiKey?: boolean;
  inventoryStatusColumnIndex?: number;
  ebayListingConditionColumnIndex?: number;
  ebayOAuthRedirectUri?: string;
  isConfigured: boolean;
  hasServiceAccount: boolean;
  /** True when Postgres schema is missing (run drizzle-kit push) */
  databaseNeedsMigration?: boolean;
  /** User-facing hint in Japanese */
  setupMessageJa?: string;
}