# lib/api-zod/src/generated/types/spreadsheetConfigRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * eBay Price Checker API
 * OpenAPI spec version: 0.1.0
 */

export interface SpreadsheetConfigRequest {
  spreadsheetId: string;
  sheetName: string;
  /** Column index (0-based) for price alert */
  alertColumnIndex?: number;
  /** Column index (0-based) for competitor price write-back */
  priceColumnIndex?: number;
  /** Column index (0-based) where my own price is stored in the sheet */
  myPriceColumnIndex?: number;
  sourceUrlColumnIndex?: number;
  ebayUrlColumnIndex?: number;
  lowestPriceColumnIndex?: number;
  lowestConditionColumnIndex?: number;
  lastCheckColumnIndex?: number;
  alertStatusColumnIndex?: number;
  repricedValueColumnIndex?: number;
  evidenceUrlsColumnIndex?: number;
  trackedTargetUrlColumnIndex?: number;
  trackedTargetConditionColumnIndex?: number;
  trackedTargetPriceColumnIndex?: number;
  autoRepriceEnabled?: boolean;
  undercutAmount?: number;
  minAllowedPrice?: number;
  ebayAppId?: string;
  ebayDevId?: string;
  ebayCertId?: string;
  ebayUserToken?: string;
  inventoryCheckerBaseUrl?: string;
  inventoryCheckerApiKey?: string;
  inventoryStatusColumnIndex?: number;
  ebayListingConditionColumnIndex?: number | null;
  ebayOAuthRedirectUri?: string;
  ebayOAuthClientSecret?: string;
  /** Google Service Account JSON (stored securely) */
  serviceAccountJson?: string;
}